+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 
+## 0.13.0
+
+- Support `h: :t` shortcut to passing `header: true`
+- Support `f: :t` shortcut to passing `footer: true`
+- Support `w: :t` shortcut to passing `wrapper: true`
+
 ## 0.12.0
 
-- Upgrade `awesome_print` to `~> 0.9.2`
+- Upgrade `awesome_print` to `~> 1.9.2`
 - Support passing pd options as part of a printed hash instead of requiring a separate hash (e.g. `pd(path: path, header: true)` instead of `pd({path: path}, header: true)` )
 - Support empty use of pd statement + options (e.g. `pd` or `pd header: true`)